generation/requirements.txt
scikit-learn
pillow
sklearn-porter